module/scope switcheroo
