+++ common/scala/pkg/src/main/scala/org/enso/pkg/Package.scala
@@ -6,7 +6,7 @@ import java.nio.file.{Files, Path}
 
 import scala.collection.JavaConverters._
 import org.apache.commons.io.FileUtils
-import org.enso.pkg.Package.QualifiedName
+import org.enso.pkg.Package.qualifiedNameSeparator
 
 import scala.io.Source
 import scala.util.Try
@@ -174,6 +174,40 @@ case class Package(root: File, config: Config) {
 }
 
 /**
+  * Represents a qualified name of a source module.
+  *
+  * @param path the names of the package and directories the module is
+  *             contained in
+  * @param module the name of the module
+  */
+case class QualifiedName(path: List[String], module: String) {
+  override def toString: String =
+    (path :+ module).mkString(qualifiedNameSeparator)
+}
+
+object QualifiedName {
+
+  /**
+    * Parses a dot-separated string representation of a qualified name into
+    * a [[QualifiedName]] object.
+    *
+    * @param qualName the string representation of a qualified name.
+    * @return the corresponding [[QualifiedName]] object.
+    */
+  def fromString(qualName: String): Option[QualifiedName] = {
+    val segments = qualName.split(qualifiedNameSeparator).toList
+    if (segments.nonEmpty) {
+      Some(QualifiedName(segments.dropRight(1), segments.last))
+    } else {
+      None
+    }
+  }
+
+  def simpleName(modName: String): QualifiedName =
+    QualifiedName(List(), modName)
+}
+
+/**
   * A companion object for static methods on the [[Package]] class.
   */
 object Package {
@@ -182,37 +216,6 @@ object Package {
   val mainFileName           = "Main.enso"
   val thumbFileName          = "thumb.png"
   val qualifiedNameSeparator = "."
-
-  /**
-    * Represents a qualified name of a source module.
-    *
-    * @param path the names of the package and directories the module is
-    *             contained in
-    * @param module the name of the module
-    */
-  case class QualifiedName(path: List[String], module: String) {
-    override def toString: String =
-      (path :+ module).mkString(qualifiedNameSeparator)
-  }
-
-  object QualifiedName {
-
-    /**
-      * Parses a dot-separated string representation of a qualified name into
-      * a [[QualifiedName]] object.
-      *
-      * @param qualName the string representation of a qualified name.
-      * @return the corresponding [[QualifiedName]] object.
-      */
-    def fromString(qualName: String): Option[QualifiedName] = {
-      val segments = qualName.split(qualifiedNameSeparator).toList
-      if (segments.nonEmpty) {
-        Some(QualifiedName(segments.dropRight(1), segments.last))
-      } else {
-        None
-      }
-    }
-  }
 
   /**
     * Creates a new Package in a given location and with config file.
